requirements.txt
anyio==4.9.0
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
cssselect2==0.8.0
distlib==0.4.0
Django==5.2.4
django-ckeditor==6.7.3
django-js-asset==3.1.2
et_xmlfile==2.0.0
exceptiongroup==1.3.0
filelock==3.18.0
firebase_admin==7.0.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
lxml==6.0.0
msgpack==1.1.1
openpyxl==3.1.5
oscrypto==1.3.0
pillow==11.3.0
platformdirs==4.3.8
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.10.1
pypdf==5.8.0
python-bidi==0.6.6
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
virtualenv==20.32.0
webencodings==0.5.1
xhtml2pdf==0.2.17
